Logic tree code generation is hard
Okay... it finally... seems to be working. Pulled a bunch of stuff out of sub-functions because gdb hates those, also because it makes what's being closed on more clear, also no forced stack scope. Doing this because "f:iv,e" produced "f:iv," even though "pinkie_pie" produced "pinkie_pie." There was a lot of code assuming the trie had null terminating sentinels, and I had to swap those for just checking if nsubs is 0. Then dump_code was an unholy nightmare that I'm already forgetting anything that I did in. It was... because the nsubs was... of the subs[i].nsubs that... ugh. Why does Spock's Brain have to be exactly how my own brain leaks knowledge?